workaround pruss pdev platform data hack only available at boot
the pruss driver expects some platform configuration data (reset management
functions) via pdev struct but this hack is only available if the platform
device exists at kernel boot time, not on devices instanciated by dt overlay.

this change adds yet another hack to instanciate the needed data in the pruss
driver if not available from the platform device.

correct fix is to make this platform data available on runtime instanciated
devices as well, or remove the need for the data at all.